third_party/WebKit/LayoutTests/external/wpt/cssom/CSSStyleRule-expected.txt

This is a testharness.js-based test.
PASS CSSRule and CSSStyleRule types 
PASS Type of CSSRule#type and constant values 
PASS Existence of CSSRule attributes 
PASS Writability of CSSRule attributes 
PASS Values of CSSRule attributes 
PASS Existence, writability and type of CSSStyleRule attributes 
PASS Values of CSSStyleRule attributes 
FAIL Mutability of CSSStyleRule's style attribute assert_equals: expected "10px" but got ""
Harness: the test ran to completion.
